No vocal rig! Just four simple 4-group Maschine rig in Ableton Live.
Lots of weird routing through Group 5, "filters," which allows me to have just two outputs from Maschine, one for Aux sends and one for Master send (Group 5 goes to master and the individual tracks have their aux sends go to the single aux out).

On each filter track there's the reverb on/off and filter (both MIDI-controlled) and each goes to both the Group -> Master and the Aux (and that send is controlled by MIDI)

**Advantages**

- Filter, 'verb and delay sends are included in lock states which are nice for morphing and all that (though I rarely use it).
- Everything is contained in Maschine, mostly, so if I needed to pop out and not use a DAW, I could easily.

**Disadvantages**

- No midi-controllable cue. This sucks and it's confirmed ![image](https://www.dropbox.com/s/887ia6jjz89ps04/Screenshot%202020-05-30%2018.41.04.png?dl=1)
- No MIDI feedback for anything. Maschine does suck this way.
- If you're using a simple MIDI controller then you need more advanced MIDI-mapping options (like being able to invert a range or limit a range). Maschine can do neither. If you're using MIDI Designer (I'm not right now), this isn't a hard limit but annoying.
- Adjusting MIDI routings is a hassle but possible.

So now I'm moving back to an older paradigm for me (but with just 4 tracks, which is new): all the MIDI-controllable parameters are going to be in Ableton, which means I need to split the kits out into four tracks of output. So that 4 tracks and no cue mix coming out of Maschine, because I'm going to cue inside Live. I would die to have Maschine **show** which tracks are on the cue bus, but that's impossible.

I even tried mapping cue to a macro control and then mapping that, which failed. So frustrating! Anyway, moving ahead with the plan. Maschine is my girlfriend but what an asshole!